"fix" release mode issue with pointer alignment
this change doesn't change anything in the binary size.
playing around with the commands in the kernel loader, sometimes
the elf loader fails on a pointer alignment assert.

if changing the kernel_bin to a static doesn't fix the issue long term,
maybe we can go back to this commit and diff the resulting executable
for further debugging.

(maybe it's a stack alignment issue?).