fix elf loader alignment issue, again
still not sure if this is actually an alignment issue, or just a
weird stack alignment bug.

if it is an issue with byte alignment, this should fix the include_bytes
step.

if this issue crops up again, we now we have to debug some more ^^
see commit 5de6549ce361a617aa8329cd5376cc223f3fbfce for reference